added a cleaned up version of sensors.ino
This is a cleaned up version of sensor.ino from multiwii. I was able to get 1500 line down to 400 lines. There is still some clean up work to do but it is a good start. It contains all the code necessary to initialize and read ACC and Gyro data from the mpu6050.